// src/app/modules/main/models/endpointheaderDto.ts
export class EndpointheaderDto {
    apiBaseFlag?:boolean;
    inputName?: string;
    inputValue?: string;
    ouputName?: string;
    outputValue?: string;
    actionType?: number;
    afterEndpointId?: number;
    status?: number;
    beforeEndpointId?: number;
    endpointDetailId?: number;
    apiId?: number;
    detailType?: number;
    checkElementPath?: string
    isSystemEndpointDetail?: number;
    apiEndpointHeaderFlag?: boolean;
}